requirements.txt
click==6.7
cffi!=1.11.3,>=1.8
cryptography>=2.1
Flask>=1.0.2
Flask-Cors>=3.0.8
Flask-RESTful>=0.3.6
itsdangerous==0.24
Jinja2>=2.10
MarkupSafe==1.0
marshmallow>=3.0.0rc8
openstacksdk>=0.31.1
keystoneauth1>=3.15.0
oslo.utils>=3.41.0
requests>=2.18.4
uWSGI==2.0.17.1
Werkzeug>=0.14.1

